+++ src/main/java/pl/thedeem/intellij/dql/services/query/DQLQuerySelectorService.java
@@ -5,6 +5,7 @@ import com.intellij.openapi.editor.Editor;
 import com.intellij.psi.PsiFile;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
+import pl.thedeem.intellij.dql.psi.DQLQuery;
 
 import java.util.function.Consumer;
 
@@ -14,6 +15,8 @@ public interface DQLQuerySelectorService {
     }
 
     @NotNull String getQueryText(@NotNull PsiFile file);
+
+    @NotNull String getQueryText(@NotNull DQLQuery query);
 
     void getQueryFromEditorContext(@NotNull PsiFile file, @Nullable Editor editor, @NotNull Consumer<@NotNull String> consumer);
 }